+++ src/components/SelectionPills/SelectionPill/docs/SelectionPill.md
@@ -4,16 +4,7 @@ SelectionPill example:
 // import { SelectionPill } from 'ic-snacks'
 import SelectionPill from '../SelectionPill.js'
 
-
 ; <div>
-    <SelectionPill
-      id="selection1"
-      isSelected={false}
-      onClick={e => {
-        console.log('SelectionPill click!', "bananas")
-      }}
-      text="bananas"
-    />
     <SelectionPill
       id="selection2"
       isSelected={true}
