+++ src/components/SelectionPills/SelectionPill/index.js
@@ -1,0 +1,3 @@
+import SelectionPill from './SelectionPill'
+
+export default SelectionPill
